requirements.txt
streamlit>=1.28.0
openai>=1.3.0
tiktoken>=0.5.1
Pillow>=10.0.0

